Moving CLI and ENV parameter functionality into parameters/ dir.
