+++ src/lib/cache/fn-cache.ts
@@ -1,6 +1,6 @@
-import { cliParams } from "../cli/arg-parser.js";
-import config from '../config/index.js';
 import log from '../log/logger.js';
+import { cliParams } from "../parameters/cli.js";
+import config from '../parameters/env.js';
 import DataDir from "./datadir.js";
 
 const cachedFns = cliParams.get('--cached-fns')?.split(',') || [];
