+++ src/lib/io/cache.ts
@@ -1,5 +1,5 @@
 import DataDir from '../cache/datadir.js';
-import config from '../config/index.js';
+import config from '../parameters/env.js';
 
 export default function cache<T>(file: string, data: T): T {
   if (!config.isProduction) {
